hooks: bootchart: start bootchart earlier
Previously, we started systemd-bootchart by adding
systemd.wants=systemd-bootchart.service to the kernel command
line. This has as a side effect that the service is added to the
default target and is started later than desired. Instead, enable the
service by default and use a new kernel command line
parameter (core.bootchart) as condition to start it.