Simplify systemd configuration for console-conf
Having `getty@.service` and `console-conf@.service` starting each others
seem to generate errors on shut down.

```
[  115.522870] systemctl[1011]: Got message type=error sender=org.freedesktop.systemd1 destination=n/a path=n/a interface=n/a member=n/a cookie=1 reply_cookie=1 signature=s error-name=org.freedesktop.systemd1.TransactionIsDestructive error-message=Transaction for getty@tty1.service/star
[  115.524325] systemctl[1011]: t is destructive (systemd-poweroff.service has 'start' job queued, but 'stop' is included in transaction).
[  115.525067] systemctl[1011]: Failed to start getty@tty1.service: Transaction for getty@tty1.service/start is destructive (systemd-poweroff.service has 'start' job queued, but 'stop' is included in transaction).
```

Instead we should just have both `console-conf@` and `getty@` enabled
and select with conditions.

Serial port services are enabled by a generator, in a similar way as
`systemd-getty-generator` for `serial-getty@`.

Automatic service for virtual terminals will not for console-conf
however.  This is because logind starts `autovt@.service` which is a
symlink to `getty@`. To have console-conf on autovt, we would need to
have `autovt@.service` start both `getty@` and `console-conf@`.
However the feature of having console-conf on all VTs is pretty
useless and does not justify a hack.

Because console-conf tries to stop itself, which it should not do, we
also disable manual stop as a work-around.