Rework disabling of unmounts
Some mounts cannot be unmounted during `shutdown.target`.  We have
tried make them part of `shutdown.target` so they are kept alive. But
because they depend on `system-systemd\\x2dfsck.slice` and
`systemd-fsck@.service` which both conflict with `shutdown.target`, it
confuses systemd, and the result of what gets scheduled is not
deterministic.

Instead we remove conflicts of `system-systemd\\x2dfsck.slice` and
`systemd-fsck@.service`. Then `DefaultDependencies=no` on mounts is
enough to disable unmounts.