Don't include secureboot-db
The secureboot-db package exists to update dbx at boot to revoke insecure
certificates/binaries.  However, changing dbx changes the measurements in
the TPM, and we seal against these measurements when doing full-disk
encryption.  So we need to NOT update dbx as a result of updating the core
snap, but instead need dbx updates to be managed more directly by snapd.