hooks: enable polkit for unconfined processes
polkit is needed in some cases for communication between services
shipped in the base, for instance when systemd-networkd talks to
systemd-hostnamed to set the hostname. Enable to fix these use cases.
Note that this does not help yet if we want to use it from a snap.

Backported from https://github.com/snapcore/core-base/pull/161