hooks: copy the desired default font to Plymouth.ttf (#307)
Previously we were moving, but now in noble Ubuntu-R.ttf was actually
a link to Ubuntu[wdth,wght].ttf, so actually a symlink to a
non-present file was created. Instead, copy the file that we have now,
which will also fix the issue if in the future it becomes a symlink
too.