static/secureboot-db.service: check mode by looking at modeenv
Checking the kernel command line to know the mode is not enough as we could be
installing UC from the initramfs. Look instead at the mode set in modeenv. Here
we need to exit with error code 2 to stop the execution as the upstream service
defines SuccessExitStatus=1.

Checking the kernel command line was causing issues as when installing from the
initramfs, the dbx was updated after keys were sealed, and the system was not
able to unseal after a reboot as dbx was not the expected one.

Eventually we should move the dbx update to make it happen in the initramfs.